6/30 11:00　更新いたしました

+++ src/main/java/com/example/mofmof/repository/entity/Tasks.java
@@ -21,7 +21,7 @@ public class Tasks {
     private String content;
 
     @Column
-    private short status;
+    private Integer status;
 
     @Column(name = "limit_date")
     private Date limitDate;
